if cross process.cmd is nil, don't kill it

+++ vm_mgr/module/core/process.go
@@ -471,6 +471,9 @@ func (p *Process) killProcess(isTxTimeout bool) {
 	for depth, process := range processDepth.processes {
 		if process != nil {
 			p.logger.Debugf("[%s] kill cross process in depth [%s]", process.processName, depth)
+			if process.cmd == nil {
+				continue
+			}
 			if err = process.cmd.Process.Kill(); err != nil {
 				p.logger.Warnf("[%s] fail to kill corss process: %s", process.processName, err)
 			}
